edit pumpValveControl to access pumps through pv_units, and able to run json progs
pump valve control is now init with only the pump_valve units and the programs. it can access the pump objects with a get function in pumpvalve class. the previously passed ser was not used so it was removed. run pump prog now can check if type is json_prog, if so executes prog_from_json with the params containg the path to the json folder and the name of that program